Made all test generators and scala-js generators to generate files only when necessary, completed this behavior now. (cherry picked from commit a996db8eb07046efac2a8068ea0dfa536e049272)

+++ project/GenLoneElement.scala
@@ -19,6 +19,8 @@ import io.Source
 import java.io.{File, FileWriter, BufferedWriter}
 
 object GenLoneElement {
+
+  val generatorSource = new File("GenLoneElement.scala")
   
   def translateLine(line: String, mapping: (String, String)*): String = {
     @tailrec
@@ -41,21 +43,23 @@ object GenLoneElement {
     def generateFile(sourceFileName: String, typeName: String, mapping: (String, String)*): File = {
       val generatedFileName = sourceFileName.replaceAll("List", typeName)
       val generatedFile = new File(targetBaseDir, generatedFileName)
-      val writer = new BufferedWriter(new FileWriter(generatedFile))
-      try {
-        val lines = Source.fromFile(new File(sourceBaseDir, sourceFileName)).getLines().toList // for 2.8
-        for (line <- lines) {
-          val generatedLine = translateLine(line, mapping.toList: _*)
-          writer.write(generatedLine.toString)
-          writer.newLine() // add for 2.8
+      if (!generatedFile.exists || generatorSource.lastModified > generatedFile.lastModified) {
+        val writer = new BufferedWriter(new FileWriter(generatedFile))
+        try {
+          val lines = Source.fromFile(new File(sourceBaseDir, sourceFileName)).getLines().toList // for 2.8
+          for (line <- lines) {
+            val generatedLine = translateLine(line, mapping.toList: _*)
+            writer.write(generatedLine.toString)
+            writer.newLine() // add for 2.8
+          }
         }
-        generatedFile
+        finally {
+          writer.flush()
+          writer.close()
+          println("Generated " + generatedFile.getAbsolutePath)
+        }
       }
-      finally {
-        writer.flush()
-        writer.close()
-        println("Generated " + generatedFile.getAbsolutePath)
-      }
+      generatedFile
     }
     
     val arrayMapping =
